requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
yara-python==4.5.1
anthropic==0.31.0
